caption/requirements.txt
Flask
Gunicorn
Pillow
opencv-python
transformers
torch